Extract post
