+++ mb-rest-api.php
@@ -31,5 +31,6 @@ if ( ! function_exists( 'mb_rest_api_load' ) ) {
 	 */
 	function mb_rest_api_load() {
 		new MetaBox\RestApi\SettingsPage;
+		new MetaBox\RestApi\Post;
 	}
 }
